refactor(game-data-codegen): extract slug helper out of the weapons module
toKebabCase was a generic string util exported from the weapons generator
only so its test could reach it. Move it to its own slug.ts (test in
slug.test.ts) so weapons.ts no longer exports it.

Kept custom rather than a library: it strips apostrophes
("Wolf's Gravestone" -> wolfs-gravestone) to match the established id
convention, whereas lodash/change-case kebab-case split on the apostrophe
and would emit "wolf-s-gravestone", changing existing ids.
