+++ tools/game-data-codegen/src/weapons.test.ts
@@ -4,23 +4,7 @@
 import { compareVersions, WEAPON_STAT_TYPES, WEAPON_TYPES } from '@genshin/game-data';
 import { describe, expect, it } from 'vitest';
 
-import { buildWeapons, toKebabCase } from './weapons.js';
-
-describe('toKebabCase', () => {
-  it('lowercases and hyphenates spaces', () => {
-    expect(toKebabCase('Mistsplitter Reforged')).toBe('mistsplitter-reforged');
-  });
-
-  it('strips apostrophes rather than hyphenating them', () => {
-    expect(toKebabCase("Wolf's Gravestone")).toBe('wolfs-gravestone');
-    expect(toKebabCase('Amos’ Bow')).toBe('amos-bow');
-  });
-
-  it('collapses runs of punctuation and trims edges', () => {
-    expect(toKebabCase('  The Catch!! ')).toBe('the-catch');
-    expect(toKebabCase('Wine & Song')).toBe('wine-song');
-  });
-});
+import { buildWeapons } from './weapons.js';
 
 describe('buildWeapons', () => {
   const weapons = buildWeapons();
